Fixes a type exception in config

+++ src/main/java/pokecube/core/blocks/repel/TileEntityRepel.java
@@ -13,7 +13,7 @@ public class TileEntityRepel extends TileEntity
 
     public TileEntityRepel()
     {
-        distance = PokecubeCore.core.getConfig().repelRadius;
+        distance = (byte) PokecubeCore.core.getConfig().repelRadius;
     }
 
     public boolean addForbiddenSpawningCoord()
